+++ examples_apps/run_tutorial_points.py
@@ -2,11 +2,15 @@
 r"""
 .. _apps-points-tutorial:
 
-Point Process Tutorial
-======================
+Point processes
+===============
 
-This tutorial shows how to generate GBS point process samples and use them to detect outlier
-points in a data set. Point processes are models for generating random point patterns and GBS
+*Technical details are available in the API documentation:* :doc:`/code/api/strawberryfields.apps.points`
+
+This section shows how to generate GBS point process samples and use them to detect outlier
+points in a data set. Point processes are models for generating random point patterns and can be
+useful in machine learning, providing a source of randomness with
+preference towards both diversity :cite:`kulesza2012determinantal` and similarity in data. GBS
 devices can be programmed to operate as special types of point processes that generate clustered
 random point patterns :cite:`jahangiri2019point`.
 
@@ -14,8 +18,8 @@ The probability of generating a specific pattern of points in GBS point processe
 matrix functions of a kernel matrix :math:`K` that describes the similarity between the points.
 Matrix functions that appear in GBS point processes are typically
 `permanents <https://en.wikipedia.org/wiki/Permanent_(mathematics)>`__ and
-`hafnians <https://the-walrus.readthedocs.io/en/latest/hafnian.html>`__. In this tutorial, we use
-the permanental point process in which the probability of observing a pattern of points :math:`S`
+`hafnians <https://the-walrus.readthedocs.io/en/latest/hafnian.html>`__. Here we use
+the permanental point process, in which the probability of observing a pattern of points :math:`S`
 depends on the permanent of their corresponding kernel submatrix :math:`K_S` as
 :cite:`jahangiri2019point`:
 
@@ -28,7 +32,7 @@ of points. Let's look at a simple example to better understand the permanental p
 
 ##############################################################################
 # We first import the modules we need. Note that the :mod:`~.apps.points` module has most of
-# the core functionalities for this tutorial.
+# the core functionalities exploring point processes.
 
 import numpy as np
 import plotly
@@ -47,7 +51,7 @@ R = np.array([(i, j) for i in range(20) for j in range(20)])
 # The rows of R are the coordinates of the points.
 #
 # Next step is to create the kernel matrix for the points of this discrete space. We call
-# the :func:`~.kernel` function which uses the *radial basis function* (RBF) kernel defined as:
+# the :func:`~.rbf_kernel` function which uses the *radial basis function* (RBF) kernel defined as:
 #
 # .. math::
 #     K_{i,j} = e^{-\|\bf{r}_i-\bf{r}_j\|^2/2\sigma^2},
@@ -70,7 +74,7 @@ R = np.array([(i, j) for i in range(20) for j in range(20)])
 #
 # Let's construct the RBF kernel with the parameter :math:`\sigma` set to 2.5.
 
-K = points.kernel(R, 2.5)
+K = points.rbf_kernel(R, 2.5)
 
 ##############################################################################
 # We generate 10 samples with an average number of 50 points per sample by calling
@@ -120,7 +124,7 @@ R = np.concatenate((clusters, noise))
 ##############################################################################
 # Then construct the kernel matrix and generate 10000 samples.
 
-K = points.kernel(R, 1.0)
+K = points.rbf_kernel(R, 1.0)
 
 samples = points.sample(K, 10.0, 10000)