+++ strawberryfields/apps/points.py
@@ -12,15 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-Point Processes
-===============
+Tools for building kernel matrices and generating point processes using GBS.
 
-**Module name:** :mod:`strawberryfields.apps.points`
+.. seealso::
 
-.. currentmodule:: strawberryfields.apps.points
-
-This module provides functions for building kernel matrices and generating point processes using
-GBS. An accompanying tutorial can be found :ref:`here <apps-points-tutorial>`.
+    :ref:`apps-points-tutorial`
 
 Point processes
 ---------------
@@ -41,15 +37,16 @@ encodes information about the similarity between points. When this kernel matrix
 semidefinite, it is possible to use GBS to implement a *permanental* point process and employ
 fast classical algorithms to sample from the resulting distribution.
 
-Summary
--------
+One choice of kernel matrix is the radial basis function (RBF) kernel whose elements are computed
+as:
 
-.. autosummary::
-    kernel
-    sample
+.. math::
+    K_{i,j} = e^{-\|\bf{r}_i-\bf{r}_j\|^2/(2\sigma^2)},
 
-Code details
-^^^^^^^^^^^^
+where :math:`\bf{r}_i` are the coordinates of point :math:`i`, :math:`\sigma` is a kernel
+parameter, and :math:`\|\cdot\|` denotes a choice of norm. The RBF kernel is positive
+semidefinite when the Euclidean norm is used, as is the case for the provided :func:`rbf_kernel`
+function.
 """
 
 import numpy as np
@@ -57,26 +54,20 @@ import scipy
 from thewalrus.csamples import generate_thermal_samples, rescale_adjacency_matrix_thermal
 
 
-def kernel(R: np.ndarray, sigma: float) -> np.ndarray:
-    r"""Calculate the kernel matrix from a set of input points.
-
-    We use the radial basis function (RBF) kernel whose elements are computed as:
-
-    .. math::
-        K_{i,j} = e^{-\|\bf{r}_i-\bf{r}_j\|^2/(2\sigma^2)},
-
-    where :math:`\bf{r}_i` are the coordinates of point :math:`i`, :math:`\sigma` is a kernel
-    parameter, and :math:`\|\cdot\|` denotes a choice of norm. The RBF kernel is positive
-    semidefinite when the Euclidean norm is used.
+def rbf_kernel(R: np.ndarray, sigma: float) -> np.ndarray:
+    r"""Calculate the RBF kernel matrix from a set of input points.
 
     The kernel parameter :math:`\sigma` is used to define the kernel scale. Points that are much
     further than :math:`\sigma` from each other lead to small entries of the kernel
     matrix, whereas points much closer than :math:`\sigma` generate large entries.
 
+    The Euclidean norm is used to measure distance in this function, resulting in a
+    positive-semidefinite kernel.
+
     **Example usage:**
 
     >>> R = np.array([[0, 1], [1, 0], [0, 0], [1, 1]])
-    >>> kernel (R, 1.0)
+    >>> rbf_kernel (R, 1.0)
     array([[1., 0.36787944, 0.60653066, 0.60653066],
            [0.36787944, 1., 0.60653066, 0.60653066],
            [0.60653066, 0.60653066, 1., 0.36787944],
@@ -87,7 +78,7 @@ def kernel(R: np.ndarray, sigma: float) -> np.ndarray:
         sigma (float): kernel parameter
 
     Returns:
-        K (array): the kernel matrix
+        K (array): the RBF kernel matrix
     """
     return np.exp(-(scipy.spatial.distance.cdist(R, R)) ** 2 / 2 / sigma ** 2)
 
@@ -95,8 +86,8 @@ def kernel(R: np.ndarray, sigma: float) -> np.ndarray:
 def sample(K: np.ndarray, n_mean: float, n_samples: int) -> list:
     """Sample subsets of points using the permanental point process.
 
-    Points can be encoded through a radial basis function kernel, provided in :func:`kernel`. Subsets
-    of points are sampled with probabilities that are proportional to the permanent of the
+    Points can be encoded through a radial basis function kernel, provided in :func:`rbf_kernel`.
+    Subsets of points are sampled with probabilities that are proportional to the permanent of the
     submatrix of the kernel selected by those points.
 
     This permanental point process is likely to sample points that are clustered together